Identity service

+++ identity-service/src/main/java/com/mycareportal/identity/service/UserService.java
@@ -2,16 +2,13 @@ package com.mycareportal.identity.service;
 
 import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import com.mycareportal.identity.constant.PredefinedRole;
 import com.mycareportal.identity.dto.request.UserCreationRequest;
 import com.mycareportal.identity.dto.request.UserUpdateRequest;
 import com.mycareportal.identity.dto.response.UserResponse;
-import com.mycareportal.identity.entity.Role;
 import com.mycareportal.identity.entity.User;
 import com.mycareportal.identity.exception.AppException;
 import com.mycareportal.identity.exception.ErrorCode;
@@ -33,30 +30,33 @@ public class UserService {
 	RoleRepository roleRepository;
 	UserMapper userMapper;
 	PasswordEncoder passwordEncoder;
-
+	
+	// create new user
 	public UserResponse createUser(UserCreationRequest request) {
 		// check if username is existed
 		boolean usernameExists = userRepository.existsByUsername(request.getUsername());
 		if (usernameExists) {
-		    throw new AppException(ErrorCode.USER_EXISTED);
+		    throw new AppException(ErrorCode.USER_ALREADY_EXISTED);
 		}
 		
 		// create new user if not existed
 		User user = userMapper.toUser(request);
 
-		Set<Role> roles = new HashSet<>();
-		roleRepository.findById(PredefinedRole.PATIENT_ROLE).ifPresent(roles::add);
-		user.setRoles(roles);
+		var roles = roleRepository.findAllById(request.getRoles());
+		user.setRoles(new HashSet<>(roles));
+		user.setPassword(passwordEncoder.encode(request.getPassword()));
 
 		user = userRepository.save(user);
 
 		return userMapper.toUserResponse(user);
 	}
-
+	
+	// get all users
 	public List<UserResponse> getUsers() {
 		return userRepository.findAll().stream().map(userMapper::toUserResponse).toList();
 	}
-
+	
+	// update a user
 	public UserResponse updateUser(String id, UserUpdateRequest request) {
 		User user = userRepository.findById(id).orElseThrow(() -> new RuntimeException("No user found"));
 		userMapper.updateUser(user, request);
@@ -67,8 +67,12 @@ public class UserService {
 
 		return userMapper.toUserResponse(userRepository.save(user));
 	}
-
+	
+	// delete a user
 	public void deleteUser(String userId) {
+		if (!userRepository.existsById(userId)) {
+			throw new AppException(ErrorCode.USER_NOT_FOUND);
+		}
 		userRepository.deleteById(userId);
 	}
 }
